Fix predict to save 32bit

+++ predict.py
@@ -9,7 +9,7 @@ from model import PedalNet
 
 
 def save(name, data):
-    wavfile.write(name, 44100, data.flatten().astype(np.int16))
+    wavfile.write(name, 44100, data.flatten().astype(np.float32))
 
 
 @torch.no_grad()
